alembic/requirements.txt
alembic==1.15.2
pydantic==2.8.2
pydantic_settings==2.8.1
psycopg2-binary==2.9.10
